AUT-5631 Moved SaaS OBBR Tests to new location.
