+++ tests/cypress/integration/saas/obuk/SmokeConsentSelfServiceTests.ts
@@ -1,17 +1,17 @@
-import {AcpLoginPage} from '../../pages/acp/AcpLoginPage';
-import {ConsentPage} from '../../pages/consent/ConsentPage';
-import {Credentials} from "../../pages/Credentials";
-import {ConsentSelfServicePage} from '../../pages/consent-self-service/ConsentSelfServicePage';
-import {Urls} from "../../pages/Urls";
-import {MfaPage} from "../../pages/mfa/MfaPage";
-import {FinancrooLoginPage} from "../../pages/financroo/FinancrooLoginPage";
-import {FinancrooWelcomePage} from "../../pages/financroo/FinancrooWelcomePage";
-import {FinancrooAccountsPage} from "../../pages/financroo/accounts/FinancrooAccountsPage";
-import {FinancrooInvestmentsPage} from "../../pages/financroo/investments/FinancrooInvestmentsPage";
-import {FinancrooContributePage} from "../../pages/financroo/investments/FinancrooContributePage";
-import {ConsentSelfServiceApplicationPage} from "../../pages/consent-self-service/ConsentSelfServiceApplicationPage";
-import {EnvironmentVariables} from "../../pages/EnvironmentVariables"
-import { FinancrooModalPage } from '../../pages/financroo/accounts/FinancrooModalPage';
+import {AcpLoginPage} from '../../../pages/acp/AcpLoginPage';
+import {ConsentPage} from '../../../pages/consent/ConsentPage';
+import {Credentials} from "../../../pages/Credentials";
+import {ConsentSelfServicePage} from '../../../pages/consent-self-service/ConsentSelfServicePage';
+import {Urls} from "../../../pages/Urls";
+import {MfaPage} from "../../../pages/mfa/MfaPage";
+import {FinancrooLoginPage} from "../../../pages/financroo/FinancrooLoginPage";
+import {FinancrooWelcomePage} from "../../../pages/financroo/FinancrooWelcomePage";
+import {FinancrooAccountsPage} from "../../../pages/financroo/accounts/FinancrooAccountsPage";
+import {FinancrooInvestmentsPage} from "../../../pages/financroo/investments/FinancrooInvestmentsPage";
+import {FinancrooContributePage} from "../../../pages/financroo/investments/FinancrooContributePage";
+import {ConsentSelfServiceApplicationPage} from "../../../pages/consent-self-service/ConsentSelfServiceApplicationPage";
+import {EnvironmentVariables} from "../../../pages/EnvironmentVariables"
+import { FinancrooModalPage } from '../../../pages/financroo/accounts/FinancrooModalPage';
 
 
 describe(`Smoke Consent self service app`, () => {
