+++ tests/cypress/integration/saas/obuk/SmokeFinancrooAccountsTests.ts
@@ -1,12 +1,12 @@
-import {AcpLoginPage} from '../../pages/acp/AcpLoginPage';
-import {FinancrooLoginPage} from '../../pages/financroo/FinancrooLoginPage';
-import {ConsentPage} from '../../pages/consent/ConsentPage';
-import {FinancrooWelcomePage} from '../../pages/financroo/FinancrooWelcomePage';
-import {FinancrooAccountsPage} from '../../pages/financroo/accounts/FinancrooAccountsPage';
-import {Credentials} from "../../pages/Credentials";
-import {Urls} from "../../pages/Urls";
-import {MfaPage} from "../../pages/mfa/MfaPage";
-import {EnvironmentVariables} from "../../pages/EnvironmentVariables"
+import {AcpLoginPage} from '../../../pages/acp/AcpLoginPage';
+import {FinancrooLoginPage} from '../../../pages/financroo/FinancrooLoginPage';
+import {ConsentPage} from '../../../pages/consent/ConsentPage';
+import {FinancrooWelcomePage} from '../../../pages/financroo/FinancrooWelcomePage';
+import {FinancrooAccountsPage} from '../../../pages/financroo/accounts/FinancrooAccountsPage';
+import {Credentials} from "../../../pages/Credentials";
+import {Urls} from "../../../pages/Urls";
+import {MfaPage} from "../../../pages/mfa/MfaPage";
+import {EnvironmentVariables} from "../../../pages/EnvironmentVariables"
 
 describe(`Smoke Financroo app`, () => {
   const acpLoginPage: AcpLoginPage = new AcpLoginPage();
